host calendar cont'd + routing
