+++ src/app/app.routes.ts
@@ -12,8 +12,10 @@ import { Trips } from './Pages/trips/trips';
 import { Wishlist } from './Pages/wishlist/wishlist';
 import { ProfileInfo } from './Pages/profile-info/profile-info';
 import { NotFound } from './Pages/not-found/not-found';
-import { HostDashboard } from './components/host/pt2/pages/host-dashboard/host-dashboard';
 import { AvailabilityCalendarComponent } from './components/host/pt2/components/availability-calendar/availability-calendar';
+import { HostCalendarPage } from './components/host/pt2/pages/host-calendar-page/host-calendar-page';
+import { HostDashboard } from './components/host/pt2/pages/host-dashboard/host-dashboard';
+
 export const routes: Routes = [
   { path: "", redirectTo:"home", pathMatch: "full"},
   { path: "home", component: Home , pathMatch:"full" },
@@ -28,8 +30,7 @@ export const routes: Routes = [
   { path: "profile", component: Profile , pathMatch:"full" },
   { path: "profileInfo", component: ProfileInfo , pathMatch:"full" },
   { path: "checkout", component: Checkout , pathMatch:"full" },
-  { path: "**", component: NotFound},
-  {path: "hostDashboard", component: HostDashboard, pathMatch: "full"},
-  {path: "hostDashboard/Calendar", component: AvailabilityCalendarComponent, pathMatch: "full"},
-
+  { path: 'host/calendar',component: HostCalendarPage },
+  { path: 'host', component: HostDashboard }, //layout shell
+  { path: "**", component: NotFound} //MUST BE AT THE END
 ];
