+++ src/app/app.ts
@@ -1,11 +1,12 @@
 import { Component } from '@angular/core';
 import { RouterOutlet } from '@angular/router';
 import { Nav } from './Shared/nav/nav';
-import { HostDashboard } from './components/host/pt2/pages/host-dashboard/host-dashboard';
+import { HostCalendarPage } from './components/host/pt2/pages/host-calendar-page/host-calendar-page';
+import { HostDashboard } from "./components/host/pt2/pages/host-dashboard/host-dashboard";
 
 @Component({
   selector: 'app-root',
-  imports: [RouterOutlet, Nav, HostDashboard],
+  imports: [RouterOutlet, Nav, HostCalendarPage, HostDashboard],
   templateUrl: './app.html',
   styleUrl: './app.css'
 })
