+++ src/app/components/host/pt2/components/availability-calendar/availability-calendar.css
@@ -11,7 +11,7 @@
 /* Calendar Section */
 .calendar-section {
   flex: 1;
-  max-width: 700px;
+  max-width: 750px;
 }
 
 .month-header {
@@ -402,8 +402,9 @@
 
 /* Settings Sidebar */
 .settings-sidebar {
-  width: 320px;
+  width: 300px;
   flex-shrink: 0;
+  margin-left: 50px;
 }
 
 .settings-section {
