+++ src/app/components/host/pt2/components/nav-host-dashboard/nav-host-dashboard.css
@@ -52,12 +52,15 @@
   color: #222222;
 }
 
-.nav-link.active {
+/* .nav-link.active {
+  color: #222222;
+  font-weight: 500;
+} */
+.active-link {
   color: #222222;
   font-weight: 500;
 }
-
-.nav-link.active::after {
+.active-link::after {
   content: '';
   position: absolute;
   bottom: -2px;
@@ -67,6 +70,16 @@
   width: 20px;
   background-color: #222222;
 }
+/* .nav-link.active::after {
+  content: '';
+  position: absolute;
+  bottom: -2px;
+  left: 35%;
+  right: 0;
+  height: 2.5px;
+  width: 20px;
+  background-color: #222222;
+} */
 
 .menu-container {
   position: relative;
