+++ src/app/components/host/pt2/components/nav-host-dashboard/nav-host-dashboard.html
@@ -7,8 +7,8 @@
 
         <!-- Navigation Links -->
         <div class="nav-links">
-          <a href="#" class="nav-link active">Today</a>
-          <a href="#" class="nav-link">Calendar</a>
+          <a href="#" class="nav-link">Today</a>
+          <a routerLink="host/calendar" routerLinkActive="active-link" class="nav-link">Calendar</a>
           <a href="#" class="nav-link">Listings</a>
           <a href="#" class="nav-link">Messages</a>
           <div class="menu-container">
