+++ src/app/components/host/pt2/pages/host-calendar-page/host-calendar-page.css
@@ -1,0 +1,79 @@
+/* Property Selector Styles */
+.property-selector {
+  margin-bottom: 24px;
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+}
+.container {
+  margin: 0 auto;
+  padding: 20px 0px;
+}
+.property-selector label {
+  font-size: 16px;
+  color: #717171;
+  line-height: 20px;
+  font-weight: 400;
+  font-family: 'Circular', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif;
+}
+
+#propertySelect {
+  padding: 8px 12px;
+  border: 1px solid #dddddd;
+  border-radius: 8px;
+  background-color: white;
+  font-size: 16px;
+  font-weight: 500;
+  color: #222222;
+  cursor: pointer;
+  appearance: none;
+  transition: border-color 0.2s ease;
+  font-family: 'Circular', -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Helvetica, Arial, sans-serif;
+  background-image: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="12" height="12" viewBox="0 0 12 12"><path fill="%23717171" d="M6 8.5L2.5 5h7L6 8.5z"/></svg>');
+  background-repeat: no-repeat;
+  background-position: right 12px center;
+  padding-right: 40px;
+}
+
+#propertySelect:hover {
+  border-color: #222222;
+}
+
+#propertySelect:focus {
+  outline: none;
+  border-color: #ff385c;
+  box-shadow: 0 0 0 2px rgba(255, 56, 92, 0.2);
+}
+
+/* Option styling (limited browser support) */
+#propertySelect option {
+  padding: 8px 12px;
+  font-size: 16px;
+  color: #222222;
+  background-color: white;
+}
+
+/* Responsive adjustments */
+@media (max-width: 768px) {
+  .property-selector {
+    margin-bottom: 20px;
+  }
+  
+  #propertySelect {
+    font-size: 15px;
+    padding: 10px 12px;
+    padding-right: 36px;
+  }
+}
+
+@media (max-width: 480px) {
+  .property-selector {
+    margin-bottom: 16px;
+  }
+  
+  #propertySelect {
+    font-size: 14px;
+    padding: 8px 10px;
+    padding-right: 32px;
+  }
+}
